requirements.txt
Django==5.0.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
pytest==8.3.4
pytest-django==4.8.0
